Add files via upload
Added support for duplicate last names iff the entry has a different phone and email e.g. relations or work phone/email. Also provided clear error messages in the case of an empty file and in delete.php where no entry can be found. Provided more informative errors in the case the user enters a phone or email which is already taken. Lastly, replaced the JSON.txt file with JSON.json as it should be.